Rearranging app layout, updating README
