+++ src/components/submit/initialDetails/fieldsets/GridTileFieldset.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Field } from 'formik';
 
-import RenderField from '../../form/RenderField';
+import RenderField from '../../../form/RenderField';
 
 /**
   Fieldset for grid tile selection (wrapper around GridTilesSelectTypeahead).
