+++ src/components/submit/surveyDetails/fieldsets/FurtherInformationFieldset.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Field } from 'formik';
 
-import RenderField from '../../form/RenderField';
+import RenderField from '../../../form/RenderField';
 
 const atLeastOneKeaObserved = values =>
   values.hours.reduce((accumulator, current) => accumulator || current.kea, false);
